refactor: move AppPickerStrategy to ui/widget package
AppPickerStrategy is a real UI-framework component class (Keep-annotated,
reflectively instantiated by SeslAppPickerGridView), not a plain
extension-function file, so it belongs alongside NoEntryView/InfoBottomSheet/
DimmingView/TouchBlockingView in ui/widget rather than ui/utils (Task 5's
extension-function moves deliberately excluded it for this reason).

Note: consumers referencing the old FQN via string (e.g. layout XML
app:strategy attributes) must be updated separately - this is a
compile-clean, runtime-breaking change for any such consumer until updated.
